refactor: 카드 컴포넌트 Card{정보} -> Card{정보}Input/Select로 네이밍 변경
