+++ src/pages/CardFormPage/CardFormPage.tsx
@@ -2,11 +2,11 @@ import styles from "./CardFormPage.module.css";
 import useCardNumbersInput from "../../hooks/useCardNumbersInput";
 import useCvcNumberInput from "../../hooks/useCvcNumberInput";
 import useExpirationDateInput from "../../hooks/useExpirationDateInput";
-import CardPassword from "./CardPassword/CardPassword";
-import CardCvcNumber from "./CardCvcNumber/CardCvcNumber";
-import CardExpirationDate from "./CardExpirationDate/CardExpirationDate";
-import CardNumber from "./CardNumber/CardNumber";
-import CardBrand from "./CardBrand/CardBrand";
+import CardPasswordInput from "./CardPasswordInput/CardPasswordInput";
+import CardCvcNumberInput from "./CardCvcNumberInput/CardCvcNumberInput";
+import CardExpirationDateInput from "./CardExpirationDateInput/CardExpirationDateInput";
+import CardNumberInput from "./CardNumberInput/CardNumberInput";
+import CardBrandSelect from "./CardBrandSelect/CardBrandSelect";
 import Button from "../../components/Button/Button";
 import { useNavigate } from "react-router-dom";
 import PreviewCardLayout from "../../components/PreviewCard/PreviewCardLayout";
@@ -59,7 +59,7 @@ export default function CardFormPage() {
       />
       <form className={styles["card-form"]}>
         {step >= CARD_STEP.PASSWORD && (
-          <CardPassword
+          <CardPasswordInput
             handleChange={onPasswordChange}
             step={step}
             handleStep={nextStep}
@@ -68,7 +68,7 @@ export default function CardFormPage() {
           />
         )}
         {step >= CARD_STEP.CVC && (
-          <CardCvcNumber
+          <CardCvcNumberInput
             handleChange={onCvcNumberChange}
             step={step}
             handleStep={nextStep}
@@ -77,7 +77,7 @@ export default function CardFormPage() {
           />
         )}
         {step >= CARD_STEP.EXPIRATION && (
-          <CardExpirationDate
+          <CardExpirationDateInput
             handleChange={onExpirationDateChange}
             step={step}
             handleStep={nextStep}
@@ -86,7 +86,7 @@ export default function CardFormPage() {
           />
         )}
         {step >= CARD_STEP.NUMBER && (
-          <CardNumber
+          <CardNumberInput
             handleChange={onCardNumberChange}
             step={step}
             handleStep={nextStep}
@@ -94,7 +94,7 @@ export default function CardFormPage() {
             errorMessage={cardNumbersError}
           />
         )}
-        <CardBrand
+        <CardBrandSelect
           handleChange={onCardBrandChange}
           step={step}
           handleStep={nextStep}
