+++ src/stories/CardCvcNumber.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CardCvcNumber from "../pages/CardFormPage/CardCvcNumber/CardCvcNumber";
+import CardCvcNumber from "../pages/CardFormPage/CardCvcNumberInput/CardCvcNumberInput";
 
 const meta: Meta<typeof CardCvcNumber> = {
   title: "Example/CardCvcNumber",
